Add information msg for wrong ssf
Error given by the os function to open the sources and sinks file is
propagated.
This error will be reported on the command line and end the program by
os.Exit(1).
So, this improves the user experiences when handling with a wrong ssf file as
in the prior version the error was ignored.